Create board and cells
This starts development by creating a board class. This class fills a mine field with cells. The cells keep track of their coordinates. I picked a coordinate system for the array of cells because I have a feeling it'll come in handy when determining the mine count for each cell.